Adds page folders
Intialized Home, Login, and SignUp pages for Surafel.
