+++ Client/src/pages/Login/Login.jsx
@@ -1,23 +1,19 @@
 import React, { useState } from 'react';
 import axios from 'axios';
+import './login.css';
 
-function Registerexample() {
+function Login() {
   const [username, setUsername] = useState('');
-  const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
-  const [rpassword, setRPassword] = useState('');
 
   const handleSubmit = async (e) => {
     e.preventDefault();
     const data = {
       username,
-      email,
-      password,
-      rpassword
-      
+      password
     };
     try {
-      const response = await axios.post('/catalog/register/', data);
+      const response = await axios.post('/catalog/login/', data);
       console.log(response.data);
     } catch (error) {
       console.error(error);
@@ -35,15 +31,7 @@ function Registerexample() {
           onChange={(e) => setUsername(e.target.value)}
         />
       </div>
-      <div>
-        <label htmlFor="email">Email:</label>
-        <input
-          type="text"
-          id="email"
-          value={email}
-          onChange={(e) => setEmail(e.target.value)}
-        />
-      </div>
+      
       <div>
         <label htmlFor="password">Password:</label>
         <input
@@ -53,18 +41,10 @@ function Registerexample() {
           onChange={(e) => setPassword(e.target.value)}
         />
       </div>
-      <div>
-        <label htmlFor="rpassword">Re-password:</label>
-        <input
-          type="text"
-          id="rpassword"
-          value={rpassword}
-          onChange={(e) => setRPassword(e.target.value)}
-        />
-      </div>
-      <button type="submit">Sign Up</button>
+
+      <button type="submit">Login</button>
     </form>
   );
 }
 
-export default Registerexample;
+export default Login;
