+++ Client/src/pages/index.js
@@ -1,5 +1,5 @@
 // export all the pages here so we can import multiple pages with one line
 // export { default as SomePage } from './SomePage';
-export { default as HomePage } from './HomePage';
-export { default as Login } from './Login';
-export { default as Registerexample } from './Registerexample';
+export { default as HomePage } from './Home/Home';
+export { default as Login } from './Login/Login';
+export { default as Registerexample } from './SignUp/SignUp';
