+++ Server/Server/urls.py
@@ -7,5 +7,6 @@ urlpatterns = [
     path("admin/", admin.site.urls),
     # path('', RedirectView.as_view(url='catalog/')),
     path('', TemplateView.as_view(template_name='index.html')),
-
+    path('login/', TemplateView.as_view(template_name='index.html')),
+    path('register/', TemplateView.as_view(template_name='index.html')),
 ]
